services/articles-service/app/requirements.txt
grpcio~=1.73.0
grpcio-tools~=1.73.0
protobuf~=6.31.1
SQLAlchemy~=2.0.29
SQLAlchemy-Serializer~=1.4.22
minio~=7.2.15
python-dotenv~=1.1.0